# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  schemas.SwaggerMovieResponse:
    properties:
      description:
        type: string
      director:
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      producer:
        type: string
      runtime:
        type: integer
      screenshots:
        items:
          type: string
        type: array
      seasons:
        type: integer
      series:
        type: integer
      similar_movies:
        items:
          $ref: '#/definitions/schemas.SwaggerMovieResponse'
        type: array
      stars:
        type: integer
      video_url:
        type: string
      watched_at:
        type: string
      year:
        type: integer
    required:
    - name
    type: object
...
  schemas.UserInput:
    properties:
      email:
        type: string
      first_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      year_of_birth:
        type: string
    type: object
  schemas.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      phone_number:
        type: string
      year_of_birth:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Ozinshe
  version: 1.0.0
paths:
...
  /user:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Successful
          schema:
            $ref: '#/definitions/schemas.UserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get user
      tags:
      - user
    put:
      consumes:
      - application/json
      parameters:
      - description: body json
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/schemas.UserInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - user
  /user/favorites:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Successful
          schema:
            $ref: '#/definitions/schemas.SwaggerMovieResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get user favorite movies
      tags:
      - user
  /user/favorites/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete user favorite movie
      tags:
      - user
    post:
      consumes:
      - application/json
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add user favorite movie
      tags:
      - user
  /user/watched-movies:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: Successful
          schema:
            $ref: '#/definitions/schemas.SwaggerMovieResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get user watched movies
      tags:
      - user
  /user/watched-movies/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete user watched movie
      tags:
      - user
    post:
      consumes:
      - application/json
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Add user watched movie
      tags:
      - user
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"